requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
Werkzeug==2.3.8
